fix(heredoc): norm
