+++ srcs/exec/expand/expand_for_heredoc.c
@@ -13,10 +13,12 @@ static bool	is_expand_in_heredoc(t_redirect *redirect)
 	t_token	*token;
 
 	token = (t_token *)redirect->tokens->node->content;
-	return (token->quote != QUOTE_NONE);
+	return (token->quote == QUOTE_NONE);
 }
 
-static t_result	expand_and_transfer_heredoc(int raw_fd, int expand_fd, t_context *context)
+static t_result	expand_and_transfer_heredoc(int raw_fd, \
+											int expand_fd, \
+											t_context *context)
 {
 	char	*line;
 	char	*expand_line;
@@ -37,31 +39,25 @@ static t_result	expand_and_transfer_heredoc(int raw_fd, int expand_fd, t_context
 	return (SUCCESS);
 }
 
-static t_result	expand_variables_in_heredoc(t_redirect *redirect, t_context *context)
+static t_result	expand_variables_in_heredoc(t_redirect *redirect, \
+											t_context *context)
 {
 	int			raw_fd;
 	int			expand_fd;
 	char		*new_filename;
 	t_result	result;
 
-	// open
 	raw_fd = open(redirect->heredoc_filename, O_RDONLY);
-	if 	(raw_fd == OPEN_ERROR)
+	if (raw_fd == OPEN_ERROR)
 		return (PROCESS_ERROR); // todo: error?
-	result = open_heredoc_fd(&expand_fd, &new_filename); // todo: func name create_
+	result = create_filename_and_open_heredoc_fd(&expand_fd, &new_filename);
 	if (result == PROCESS_ERROR)
 		return (PROCESS_ERROR); // todo: error?
-
-	// expand: read from before_fd -> expand -> write to after_fd
 	result = expand_and_transfer_heredoc(raw_fd, expand_fd, context);
 	if (result == PROCESS_ERROR)
 		return (PROCESS_ERROR); // todo: error?
-
-	// close
 	close(raw_fd); // todo: error
 	close(expand_fd); // todo: error
-
-	// replace before->after
 	unlink(redirect->heredoc_filename); // todo: error
 	ft_free(&redirect->heredoc_filename);
 	redirect->heredoc_filename = new_filename;
